Bluetooth: Mesh: Add support feature for proxy client
1. Split proxy.c to proxy_common.c and proxy_server.c
2. Add Proxy Cient functionality to proxy_client.c

Add support for scan and parse network_id or node_id
   beacon, and connect to proxy server. discovery servers.

Move Mesh GATT Services to separate files, just like Chapter
6 and Chapter 7 of the Mesh Specification.

Add support provisioner over pb-gatt in the future.
