+++ subsys/bluetooth/mesh/cdb.c
@@ -7,6 +7,7 @@
 #include <zephyr.h>
 #include <string.h>
 
+#include <bluetooth/bluetooth.h>
 #include <bluetooth/mesh.h>
 
 #define BT_DBG_ENABLED IS_ENABLED(CONFIG_BT_MESH_DEBUG_CDB)
