+++ subsys/bluetooth/mesh/proxy_client.h
@@ -1,0 +1,11 @@
+/*  Bluetooth Mesh */
+
+/*
+ * Copyright (c) 2020 Lingao Meng
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+void bt_mesh_proxy_client_process(const bt_addr_le_t *addr, int8_t rssi,
+				  struct net_buf_simple *buf);
+int bt_mesh_proxy_client_init(void);
